level1 - changed animations to extension functions.
