--- app/src/main/java/com/project/myapp/ext/Context.kt
@@ -1,8 +1,0 @@
-package com.project.myapp.ext
-
-import android.content.Context
-import android.widget.Toast
-
-fun Context.toast(message: String) {
-    Toast.makeText(this, message, Toast.LENGTH_SHORT).show()
-}
