+++ app/src/main/java/com/project/myapp/screens/main/MainActivity.kt
@@ -3,12 +3,11 @@ package com.project.myapp.screens.main
 import android.content.Intent
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
-import androidx.core.app.ActivityOptionsCompat
 import com.project.myapp.KeysHolder.USER_NAME_KEY
-import com.project.myapp.R
 import com.project.myapp.databinding.ActivityMainBinding
 import com.project.myapp.ext.componentactivity.handleBackPress
 import com.project.myapp.ext.componentactivity.setEnableEdgeToEdge
+import com.project.myapp.ext.context.customAnimationBackward
 import com.project.myapp.ext.view.setVisualize
 import com.project.myapp.screens.auth.AuthActivity
 
@@ -38,14 +37,8 @@ class MainActivity : AppCompatActivity() {
     }
 
     private fun goToPreviousActivity() {
-        val animation =
-            ActivityOptionsCompat.makeCustomAnimation(
-                this@MainActivity,
-                R.anim.slide_in_right,
-                R.anim.slide_out_right,
-            )
         val intent = Intent(this@MainActivity, AuthActivity::class.java)
-        startActivity(intent, animation.toBundle())
+        startActivity(intent, customAnimationBackward().toBundle())
         finish()
     }
